LayoutTests/fast/flexbox/box-orient-button-expected.txt

Check if box-orient is working in <button>. See https://bugs.webkit.org/show_bug.cgi?id=25406

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gebi('default').clientHeight is referenceHorizontalHeight
PASS gebi('horizontal').clientHeight is referenceHorizontalHeight
PASS gebi('vertical').clientHeight is referenceVerticalHeight
PASS gebi('toHorizontal').clientHeight is referenceHorizontalHeight
PASS gebi('toVertical').clientHeight is referenceVerticalHeight
PASS successfullyParsed is true

TEST COMPLETE
